refactor: Zeichnen der Milestones, Tasks und X-Achse aus chart.components.ts herausgelöst fix: DIPA-106 Tooltip & Sprungkorrektur
