+++ frontend/projects/dipa-frontend/src/app/modules/gantt/chart/chart-elements/IChartElement.ts
@@ -1,0 +1,14 @@
+export interface IChartElement {
+
+  svg: any;
+  xScale: any;
+  data: any[];
+  tooltip: any;
+
+  dateOptions: any;
+
+  draw(offset): void;
+  redraw(offset): void;
+  showTooltip(d, x, y): void;
+  getAreaHeight(): number;
+}
